refactor : server - app.js 코드 분리
- server - config - database.js : createPool 코드를 넣어두었습니다.
- server - router : app.js에서 분리해둔 api 코드들을 넣어두었습니다.
- server - util - dbquery.js : createActiveLog 함수가 공통적으로 사용되기 때문에 util이라는 폴더 내에 따로 분리해뒀습니다.
